feat: create upload images with Cloudnary

+++ pages/api/products/[id].js
@@ -2,7 +2,9 @@ import dbConect from "../../../util/mongo";
 import Product from "../../../models/Product"
 
 export default async function handler(req,res){
-  const { method, query:{id} } = req;
+  const { method, query:{id}, cookies } = req;
+
+  const token = cookies.token;
 
   await dbConect();
 
@@ -17,6 +19,9 @@ export default async function handler(req,res){
   }
   
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN){
+      return res.status(401).json("Not authenticated!")     
+    }
     try {
       const product = await Product.create(req.body);
       res.status(201).json(product)
@@ -28,6 +33,9 @@ export default async function handler(req,res){
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN){
+      return res.status(401).json("Not authenticated!")     
+    }
     try {
       await Product.findByIdAndDelete(id)
       res.status(200).json({message:"O produto foi excluído"})
